Change Money fields to Decimal in Rails Admin (#588)
* Change Rails Admin Money field to extend String

This allows for editing a money field with decimals in it.

* Change Rails Admin Money field to extend Decimal

+++ lib/money-rails/rails_admin.rb
@@ -8,7 +8,7 @@ module RailsAdmin
   module Config
     module Fields
       module Types
-        class Money < RailsAdmin::Config::Fields::Types::Integer
+        class Money < RailsAdmin::Config::Fields::Types::Decimal
           RailsAdmin::Config::Fields::Types::register(self)
 
           register_instance_option :pretty_value do
